Create SbSolution.scala
Program parses the test datafile and creates and identifies the keywords file useful for for business problem.
Once the keyword file is reviewed and finalized it is provided as input to the program as second argument.  Also passed to the program as argument 1 the log files for processing.  The log records are processed and in memory Maps are maintained for keyword->Ip 
Also, various files are created for reporting purposes as described in README file